+++ integration_tests/test_ica.py
@@ -1,8 +1,6 @@
-import base64
 import json
 
 import pytest
-from cprotobuf import Field, ProtoEntity
 from pystarport import cluster
 
 from .cosmoscli import module_address
@@ -11,6 +9,7 @@ from .ibc_utils import (
     Status,
     deploy_contract,
     funds_ica,
+    gen_query_balance_packet,
     gen_send_msg,
     ica_send_tx,
     parse_events_rpc,
@@ -169,26 +168,8 @@ def test_ica(ibc, order, tmp_path):
     call_module_safe_query(cli_controller, connid, signer, ica_address)
 
 
-class QueryBalanceRequest(ProtoEntity):
-    address = Field("string", 1)
-    denom = Field("string", 2)
-
-
 def call_module_safe_query(cli, connid, signer, ica_address):
-    query = QueryBalanceRequest(address=ica_address, denom="basecro")
-    data = json.dumps(
-        {
-            "@type": "/ibc.applications.interchain_accounts.host.v1.MsgModuleQuerySafe",
-            "signer": ica_address,
-            "requests": [
-                {
-                    "path": "/cosmos.bank.v1beta1.Query/Balance",
-                    "data": base64.b64encode(query.SerializeToString()).decode("utf-8"),
-                }
-            ],
-        }
-    )
-    packet = cli.ica_generate_packet_data(data)
+    packet = gen_query_balance_packet(cli, ica_address)
     timeout = 60  # in seconds
     rsp = cli.ica_send_tx(
         connid,
@@ -198,6 +179,3 @@ def call_module_safe_query(cli, connid, signer, ica_address):
         from_=signer,
     )
     assert rsp["code"] == 0, rsp
-    events = parse_events_rpc(rsp["events"])
-    seq = int(events.get("send_packet")["packet_sequence"])
-    return seq
